requirements.txt
scrapy
psycopg2
python-dotenv
chompjs
openai
pgvector